pyproject.toml
[tool.poetry]
name = "electricitybot"
version = "1.1.0"
description = "Telegram electricity bot"
authors = ["Maksym Biloborodko <f1ashhimself@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.11.1"
pytz = "2025.2"
python-telegram-bot = "22.6"
pydantic-settings = "2.12.0"
matplotlib = "3.10.8"
python-dotenv = "1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pyhamcrest = "^2.1.0"
freezegun = "1.5.5"
flake8 = "7.3.0"
black = "26.1.0"
isort = "7.0.0"
pytest-cov = "7.0.0"
pytest-dotenv = "0.5.2"
flake8-pyproject = "^1.2.4"

[tool.poetry.scripts]
electricitybot = "electricitybot.bot:run_bot"

[tool.black]
line-length = 120
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.hg
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
combine_as_imports = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3
order_by_type = false
known_local_folder = ["src"]

[tool.flake8]
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "build",
    "dist",
    "venv",
    ".venv",
]
max-complexity = 10
max-line-length = 120

[tool.coverage.run]
concurrency = [
    "thread",
]

[tool.coverage.report]
precision = 2

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --tb=short"
testpaths = ["tests"]
env_override_existing_values = 1
env_files = [
    "envs/test.env",
    "envs/.test.env"
]
log_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s: %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
